requirements.txt
# Core Dependencies - Compatible with Python 3.10
numpy>=1.23.5,<2.0.0

# Vector Databases - For Python 3.10
faiss-cpu>=1.7.3
chromadb==0.4.18  # This version works with Python 3.10
pinecone>=2.2.1  # Updated from pinecone-client

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.64.0

# Choose ONE of the following sections based on your needs:

# Option 1: If you need the older OpenAI API (uncomment these lines)
# openai==0.28.0
# langchain==0.0.267  # Older version compatible with openai 0.28.0

# Option 2: If you need the newer LangChain features (uncomment these lines)
openai>=1.68.2
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-core>=0.0.15

# Optional Dependencies
sentence-transformers>=2.2.2
